+++ pkg/streamingpromql/operators/aggregations/count_values.go
@@ -276,6 +276,10 @@ func (c *CountValues) Finalize(ctx context.Context) error {
 	return c.LabelName.Finalize(ctx)
 }
 
+func (c *CountValues) Stats(ctx context.Context) (*types.OperatorEvaluationStats, error) {
+	return types.CombineStats[types.StatsProvider](ctx, c.Inner, c.LabelName)
+}
+
 func (c *CountValues) Close() {
 	c.Inner.Close()
 	c.LabelName.Close()