+++ pkg/streamingpromql/operators/deduplicate_and_merge.go
@@ -160,6 +160,10 @@ func (d *DeduplicateAndMerge) Finalize(ctx context.Context) error {
 	return d.Inner.Finalize(ctx)
 }
 
+func (d *DeduplicateAndMerge) Stats(ctx context.Context) (*types.OperatorEvaluationStats, error) {
+	return d.Inner.Stats(ctx)
+}
+
 func (d *DeduplicateAndMerge) Close() {
 	d.Inner.Close()
 }